refactor: Use useRequest to close my jio instead of reducer since it does not update state
