+++ frontend/src/pages/dashboard/jios/MyJioCard.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { useState } from 'react';
 import { Button, Card, CardHeader, Stack, Typography } from '@mui/material';
 import Iconify from '../../../components/Iconify';
 import { IconStyle } from '../../../sections/@dashboard/user/profile/common';
@@ -8,9 +9,9 @@ import { formatStartEndDate } from '../../../utils/formatTime';
 import { getPassesText } from './common';
 import CloseMyJioDialog from './CloseMyJioDialog';
 import { useSnackbar } from 'notistack';
-import { closeMyJio } from 'src/store/reducers/myJios';
 import useRefresh from 'src/hooks/useRefresh';
-import { useDispatch } from 'src/store';
+import { closeMyJio } from 'src/services/myJios';
+import { useRequest } from 'ahooks';
 
 interface MyJioCardProps {
   data: Jio;
@@ -19,8 +20,14 @@ interface MyJioCardProps {
 export default function MyJioCard({ data }: MyJioCardProps) {
   const { enqueueSnackbar } = useSnackbar();
   const refresh = useRefresh();
-  const dispatch = useDispatch();
-  const [isCloseDialogOpen, setIsCloseDialogOpen] = React.useState(false);
+  const [isCloseDialogOpen, setIsCloseDialogOpen] = useState(false);
+  const { run: submitCloseMyJio } = useRequest(closeMyJio, {
+    manual: true,
+    onSuccess: () => {
+      enqueueSnackbar('Closed!');
+      refresh();
+    },
+  });
 
   const handleClose = () => {
     setIsCloseDialogOpen(true);
@@ -30,16 +37,9 @@ export default function MyJioCard({ data }: MyJioCardProps) {
     setIsCloseDialogOpen(false);
   };
 
-  const handleDialogClose = async () => {
+  const handleDialogClose = () => {
     setIsCloseDialogOpen(false);
-    await dispatch(
-      closeMyJio(data.id, {
-        onSuccess: () => {
-          enqueueSnackbar('Closed!');
-        },
-      })
-    );
-    refresh();
+    submitCloseMyJio(data.id);
   };
 
   return (
